implement replace command
the `replace` command will do a match in the same fashion as `grep` but instead of presenting results with search patterns highlighted, it will highlight matches in red followed immediately by replacement string in green. in this way, it is easy to identify on one line what will be matched and replaced. by default, the operation will ask for confirmation before writing back to vault. it is possible to use flags `-y` to write without confirmation, `-n` to skip confirmation and not write aka dry-run, and limit scope to keys (`-k`) and values (`-v`) in the same way as in `grep`.